Update file structure and separate classes from main executable
