+++ lib/Player.rb
@@ -1,0 +1,6 @@
+class Player
+  attr_accessor :name
+  def initialize(name)
+    @name = name
+  end
+end
